reduce buffer if capacity >> required amount
Currently buffer is expand-only. For long-lived connections the buffer
will expand up to limit negotiated by @@max_allowed_packet setting
between client and server. Default is usually 16MB, which allows rows
of up to 16MB to be transferred. This is fine and necessary to support
but memory usage of the client is effectively the high watermark of
the largest row transferred over lifetime of the collection.

For an environment with many connections (e.g. pooled, many shards)
memory can be a constraint & problem.

This implementation uses an aggressive approach to downsize the buffer
by checking if current buffer capacity significantly exceeds memory
required to process latest query